requirements.txt
nose
rednose
coverage
Pillow
git+git://github.com/htryppcook/python_toolbox.git